* dia/forms.pyc: Associação funcionando! Consumo funcionando! A lua brilha lá fora!

+++ dia/models.py
@@ -17,8 +17,8 @@ class Dia(models.Model):
     data = models.DateField(blank=True,null=True)
     dias = models.CommaSeparatedIntegerField(max_length = 14)
     tipo = models.CharField(choices = TIPO_DIA,max_length = 30)
-    consumo = models.IntegerField()
-    
+    consumo = models.DecimalField(max_digits = 10,decimal_places=2)
+
     class Meta:
         ordering = ['-data']
